// Client/src/interfaces/actionTypes.ts
export enum ActionTypes {
    ADD_CHARACTER = "ADD_CHARACTER",
    DEL_CHARACTER = "DEL_CHARACTER",
    SEARCH_CHARACTER = "SEARCH_CHARACTER",
    FILTER = 'FILTER',
    ORDER = 'ORDER',
    RESET = 'RESET',
    ADD_FAV = 'ADD_FAV',
    REMOVE_FAV = 'REMOVE_FAV',
    GET_FAVORITES = "GET_FAVORITES",
    REGISTER = 'REGISTER',
    LOGIN = "LOGIN",
    ERROR_CHARACTER = "ERROR_CHARACTER",
    ERROR_USER = "ERROR_USER"
}